registry/storage: Fix GCS client authorization with non-default credentials
Previously we were creating a GCS client with the default application
credentials. Now we need to use the OAuth Token Source to ensure
we have access.

Also fix error handling of Delete response. The GCS integration tests
were failing because `storage.Delete` returned
`storage.ErrObjectNotExist` if the file were not found, and we need to
check that error instead of casting it to `googleapi.Error`.