Add GitLab documentation
This commit add a new directory, `docs-gitlab/`, to contain
documentation that is specific to the GitLab fork. This will
enable us to add documentation without increating the difficulty of
pulling in changes from upstream.

The only major content added in this commit is a guide to local
integration tests for the S3 storage driver.