Improve GC sweep logs
After the GC mark improvements, we now have duplicated "blob eligible
for deletion" and "manifest eligible for deletion" log lines. Since it's
useful to log the objects being deleted in the sweep stage (as we can
display their full path, which we can't during the mark stage) we should
rename the sweep stage logs from "... eligible for deletion" to
"preparing to delete ..." so that these are not confused.