Optimize the garbage collector sweep phase for GCS
Leverages the work done on
https://gitlab.com/gitlab-org/container-registry/-/merge_requests/24 to
implement concurrent blob and manifest deletion for the GCS driver,
using a goroutine per object to delete.

The number of concurrent delete requests is limited to 1000, as that is
the initial write limit for objects, as advertised by Google
(https://cloud.google.com/storage/quotas#objects).